Fixup metadata values that Nextclade can't parse
Replaces integer values in the metadata (0 and 1) with string versions
of the same information to work around an error thrown by Nextclade when
it tries to parse an Auspice JSON that has an integer where it expects a
string. The primary issue could be that augur export loads this column
as an integer and exports it in JSON as an integer even though the
Auspice config JSON specifies the type for the column as "categorical".
Separately, we may want `augur merge` to provide a way to output a
non-integer output column?